Add binary file data type Add hashing primitives Track node exposition for hash count Complete 2-3 tree binary factory Add root hash for 2-3 tree Add hash counting for 2-3 tree Add size for 2-3 tree Use flag for command-line options

+++ cairo-bptree/graph.go
@@ -58,7 +58,7 @@ func (g *Node23Graph) saveDot(filename string, debug bool) {
 			if n.nextKey() == nil {
 				next = "nil"
 			} else {
-				next = strconv.FormatUint(*n.nextKey(), 10)
+				next = strconv.FormatUint(uint64(*n.nextKey()), 10)
 			}
 			if debug {
 				nodeId = fmt.Sprintf("k=%v %s-%v", ptr2pte(n.keys[:len(n.keys)-1]), next, n.keys)
